TP5 restart
Tp5 code with new TP6 code commented out.  This will be used as a basis
for the new TP6 code which solves a bunch of issue but is primarily to
take into account the angular measurement which is affected  by
acceleration.  This cause a speedup in direction reverses and going
slower when turning in one direction and faster when turning the other.